Support both single-column and desktop styles.
If you add both the single-column.css and desktop.css styles at the same time in an app like Fluid they conflict and break each other.  Wrapping the desktop.css in a media query to only apply at larger widths resolves the conflicts and allows for a better experience if you resize the window from small to big (or vice versa).